Add tests for the three new lsm subsystem types
throttle_test.go: fresh tests for *writeThrottle covering state
transitions, pressure/rate clamping, callback firing only on
transition, and nil-receiver guards.

flush_pool_test.go: fresh tests for *flushPool covering Submit/Drain
through the install callback, immutables-list trim on success,
immutables-list preserved on install error, nil-shard guard via
ErrFlushNilMemtable, queue stats on completion, and nil-receiver
guards. Uses a stubMemIndex so tests can build memTables without
the full WAL/index stack.

compactor_test.go: migrates existing tests for executor methods +
compactDef + pacer-bypass from compaction_test.go and
level_manager_test.go (where they used to live by accident of the
methods having been on *levelManager). compaction_test.go shrinks to
just scheduler/policy + plan.State tests.
